+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.4 *(not released)*
 * Possible to ignoreExitValue _(konrad-garus)_
 * Added support for exec taks overrides (delegates down to Gradle exec task) _(konrad-garus)_
 * Now adding npmInstall outside afterEvaluate to allow better dependsOn usage
+* Reworked SetupTask so that it is using task input/output change tracking
 
 Version 0.3 *(2014-01-10)*
 --------------------------